backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
email-validator
httpx
